Added Login Page:
Added a minimalistic login page with soft Green color scheme,
sourced with HTML5, Javascript (Jquery), CSS, and SCSS.  The form was
designed so that even when you're typing in the form, you can see the form
label.  Also added a nice menu tool which can be used in the future.